auth_module/pyproject.toml
[project]
name = "auth-module"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "psycopg[binary,pool]>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
]
